+++ src/common/capabilities.ts
@@ -1,5 +1,27 @@
+// Return true if this is a browser environment, false if this is
+// a server side environment (node.js) or web worker.
+export function isBrowser(): boolean {
+  return 'window' in globalThis && 'document' in globalThis;
+}
+
+export function throwIfNotInBrowser(): void {
+  if (!isBrowser()) {
+    throw new Error(
+      '<math-field> is an interactive component that needs to run in a browser environment\n' +
+        'If you are using nextjs, see https://nextjs.org/docs/advanced-features/dynamic-import#with-no-ssr'
+    );
+  }
+}
+
 export function isTouchCapable(): boolean {
-  return window.matchMedia?.('(any-pointer: coarse)').matches ?? false;
+  return (
+    (isBrowser() && window.matchMedia?.('(any-pointer: coarse)').matches) ??
+    false
+  );
+}
+
+export function canVibrate(): boolean {
+  return isBrowser() && typeof navigator.vibrate === 'function';
 }
 
 export function osPlatform():
@@ -31,7 +53,7 @@ export function osPlatform():
 }
 
 export function supportRegexPropertyEscape(): boolean {
-  if (navigator === undefined) return true;
+  if (!isBrowser()) return true;
 
   if (/firefox/i.test(navigator.userAgent)) {
     const m = navigator.userAgent.match(/firefox\/(\d+)/i);
@@ -55,5 +77,5 @@ export function supportLocalFontEnumeration(): boolean {
   // Firefox and Safari return true for fonts that are not loaded...
   // https://bugzilla.mozilla.org/show_bug.cgi?id=1252821 🤦‍♂️
   // So, if on Firefox, always assume that the fonts are not loaded.
-  return !/firefox|safari/i.test(navigator.userAgent);
+  return isBrowser() && !/firefox|safari/i.test(navigator.userAgent);
 }