+++ src/editor/popover.ts
@@ -19,6 +19,7 @@ import { getCaretPoint } from '../editor-mathfield/utils';
 import type { MathfieldPrivate } from '../editor-mathfield/mathfield-private';
 import { typeset } from '../core/typeset';
 import { getDefaultRegisters } from '../core/registers';
+import { throwIfNotInBrowser } from '../common/capabilities';
 
 // A textual description of a LaTeX command.
 // The value can be either a single string, or an array of string
@@ -401,7 +402,7 @@ export function updatePopoverPosition(
     // Call ourselves again later, typically after the
     // rendering/layout of the DOM has been completed
     // (don't do it on next frame, it might be too soon)
-    window.setTimeout(() => updatePopoverPosition(mf), 100);
+    setTimeout(() => updatePopoverPosition(mf), 100);
     return;
   }
 
@@ -418,6 +419,7 @@ function setPopoverPosition(
   mf: MathfieldPrivate,
   position: { x: number; y: number; height: number }
 ): void {
+  throwIfNotInBrowser();
   // Get screen width & height (browser compatibility)
   const screenHeight =
     window.innerHeight ||